+++ package.json
@@ -15,6 +15,7 @@
     "next": "13.2.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-icons": "^4.8.0",
     "sharp": "^0.31.3",
     "swr": "^2.1.3"
   },
