--- styles/GNB.module.scss
@@ -1,90 +1,0 @@
-@import 'var';
-@import 'mixin';
-
-.blank {
-  height: 3.5rem;
-}
-
-.nav {
-  margin: 0;
-  padding: 0;
-  top: 0px;
-  left: 0px;
-  z-index: 100;
-  position: fixed;
-  @include transition(top, 0.3s, linear);
-  width: 100%;
-  height: 3.5rem;
-  background-color: map-get($colors, main);
-  @include box-shadow(true);
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  min-width: 43.75rem;
-
-  @include max-width($mixin-large-width) {
-    min-width: auto;
-    top: 0 !important;
-  }
-
-  .internet_link {
-    margin-left: 1rem;
-    padding: 0.3rem 0.5rem;
-    a {
-      img {
-        vertical-align: middle;
-      }
-    }
-
-    @include min-width($mixin-large-width) {
-      margin-left: 1.5rem;
-    }
-    @include min-width($mixin-xl-width) {
-      margin-left: 2rem;
-    }
-    @include max-width($mixin-middle-width) {
-      margin-left: 0.4rem;
-    }
-  }
-
-  .title {
-    font-size: 2rem;
-    color: white;
-    font-weight: 600;
-  }
-
-  .form_link {
-    margin-right: 1rem;
-    padding: 0.4rem 0.8rem;
-    border-radius: 10px;
-    font-size: 1rem;
-    line-height: 1.2rem;
-    color: white;
-    font-weight: 600;
-    background-color: salmon;
-    @include transition(background-color, 80ms, $bezier-curve);
-    @include box-shadow();
-
-    &:hover {
-      background-color: rgb(224, 110, 97);
-    }
-
-    &:active {
-      background-color: rgb(180, 80, 69);
-    }
-
-    a:focus-visible {
-      outline-offset: 0.7rem;
-    }
-
-    @include min-width($mixin-large-width) {
-      margin-right: 1.5rem;
-    }
-    @include min-width($mixin-xl-width) {
-      margin-right: 2rem;
-    }
-    @include max-width($mixin-middle-width) {
-      margin-right: 0.4rem;
-    }
-  }
-}
