+++ styles/_var.scss
@@ -7,5 +7,6 @@ $mixin-1440p-width: 160rem;
 $colors: (
   main: #ffc1cc,
   secondary: #fa923f,
+  salmon: #fa8072,
 );
 $bezier-curve: cubic-bezier(0.4, 0, 0.2, 1);
